Refine examples, move common code to separate file.
Removing boilerplate helps to emphasize actual library code in the
examples.
